propagate pagebreak css in Word table cells to div, p: https://github.com/metanorma/metanorma-iso/issues/888
